requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Login==0.4.0
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
python-firebase==1.2
requests==2.18.4
urllib3==1.22
Werkzeug==0.12.2
WTForms==2.1
